[agent] refactor(assembly): extract class_map + ner modules from lib.rs
